RAG/resume/requirement.txt

google-genai  
chromadb
google-api-core
streamlit